+++ src/app/host_crud.rs
@@ -103,6 +103,7 @@ impl App {
         new_host.proxy_jump = host.proxy_jump.clone();
         new_host.forward_agent = host.forward_agent.unwrap_or(false);
         new_host.remote_command = host.remote_command.clone();
+        new_host.session_logging = meta.session_logging;
         new_host.identity_id = self.match_identity_for_ssh_host(host)?;
         self.store.create_host(&new_host)?;
         Ok(name)